style: BowlingGame -> Frames, BowlingGameRunner -> BowlingGame으로 rename
